New Cheese Database Sorter and Interpreter Program
Two new classes were added. The CheeseAnalyzer Class reads in a csv file containing a database of cheeses using the fileReader() method and sorts it's datapoints via comma and quotation mark separation. The CheeseAnalyzer Class also outputs to an output.txt file (with the fileWriter() method) data interpreted by the CheeseCalculator class. The CheeseCalculator class does several comparisons to calculate the amount of specific types of data in the cheese database.